requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
DateTime==4.3
idna==3.2
logging==0.4.9.6
python-decouple==3.4
pytz==2021.1
requests==2.26.0
urllib3==1.26.6
wincertstore==0.2
zope.interface==5.4.0
